requirements.txt
python-telegram-bot
vk-api
openai
opencv-python
pandas
ultralytics
supervision
python-dotenv
tqdm